+++ src/snowflake/cli/plugins/spcs/compute_pool/manager.py
@@ -1,8 +1,13 @@
 from typing import Optional
 
+from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
-from snowflake.cli.plugins.spcs.common import strip_empty_lines
+from snowflake.cli.plugins.spcs.common import (
+    handle_object_already_exists,
+    strip_empty_lines,
+)
 from snowflake.connector.cursor import SnowflakeCursor
+from snowflake.connector.errors import ProgrammingError
 
 
 class ComputePoolManager(SqlExecutionMixin):
@@ -28,7 +33,17 @@ class ComputePoolManager(SqlExecutionMixin):
             """.splitlines()
         if comment:
             query.append(f"COMMENT = {comment}")
-        return self._execute_query(strip_empty_lines(query))
+
+        try:
+            return self._execute_query(strip_empty_lines(query))
+        except ProgrammingError as e:
+            handle_object_already_exists(e, ObjectType.COMPUTE_POOL, pool_name)
 
     def stop(self, pool_name: str) -> SnowflakeCursor:
-        return self._execute_query(f"alter compute pool {pool_name} stop all;")
+        return self._execute_query(f"alter compute pool {pool_name} stop all")
+
+    def suspend(self, pool_name: str) -> SnowflakeCursor:
+        return self._execute_query(f"alter compute pool {pool_name} suspend")
+
+    def resume(self, pool_name: str) -> SnowflakeCursor:
+        return self._execute_query(f"alter compute pool {pool_name} resume")