+++ src/snowflake/cli/plugins/spcs/image_registry/manager.py
@@ -1,10 +1,17 @@
 import base64
 import json
+import re
 from urllib.parse import urlparse
 
 import requests
 from click import ClickException
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
+from snowflake.connector.cursor import DictCursor
+
+
+class NoImageRepositoriesFoundError(ClickException):
+    def __init__(self, msg: str = "No image repository found."):
+        super().__init__(msg)
 
 
 class RegistryManager(SqlExecutionMixin):
@@ -45,3 +52,17 @@ class RegistryManager(SqlExecutionMixin):
         if resp.status_code != 200:
             raise ClickException(f"Failed to login to the repository {resp.text}")
         return json.loads(resp.text)["token"]
+
+    def _has_url_scheme(self, url: str):
+        return re.fullmatch(r"^.*//.+", url) is not None
+
+    def get_registry_url(self):
+        repositories_query = "show image repositories in account"
+        result_set = self._execute_query(repositories_query, cursor_class=DictCursor)
+        results = result_set.fetchall()
+        if len(results) == 0:
+            raise NoImageRepositoriesFoundError()
+        sample_repository_url = results[0]["repository_url"]
+        if not self._has_url_scheme(sample_repository_url):
+            sample_repository_url = f"//{sample_repository_url}"
+        return urlparse(sample_repository_url).netloc