+++ src/snowflake/cli/plugins/spcs/image_repository/manager.py
@@ -1,8 +1,16 @@
+from typing import Dict
 from urllib.parse import urlparse
 
 from click import ClickException
+from snowflake.cli.api.constants import ObjectType
+from snowflake.cli.api.project.util import (
+    escape_like_pattern,
+    is_valid_unquoted_identifier,
+)
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
-from snowflake.connector.cursor import SnowflakeCursor
+from snowflake.cli.plugins.spcs.common import handle_object_already_exists
+from snowflake.connector.cursor import DictCursor
+from snowflake.connector.errors import ProgrammingError
 
 
 class ImageRepositoryManager(SqlExecutionMixin):
@@ -15,38 +23,44 @@ class ImageRepositoryManager(SqlExecutionMixin):
     def get_role(self):
         return self._conn.role
 
-    def get_repository_url_list(self, repo_name: str) -> SnowflakeCursor:
-        role = self.get_role()
-        database = self.get_database()
-        schema = self.get_schema()
+    def get_repository_row(self, repo_name: str) -> Dict:
+        if not is_valid_unquoted_identifier(repo_name):
+            raise ValueError(
+                f"repo_name '{repo_name}' is not a valid unquoted Snowflake identifier"
+            )
 
-        registry_query = f"""
-            use role {role};
-            use database {database};
-            use schema {schema};
-            show image repositories like '{repo_name}';
-            """
+        repo_name = repo_name.upper()
 
-        return self._execute_query(registry_query)
+        # because image repositories only support unquoted identifiers, SHOW LIKE should only return one or zero rows
+        repository_list_query = (
+            f"show image repositories like '{escape_like_pattern(repo_name)}'"
+        )
 
-    def get_repository_url(self, repo_name):
-        database = self.get_database()
-        schema = self.get_schema()
-
-        result_set = self.get_repository_url_list(repo_name=repo_name)
-
+        result_set = self._execute_schema_query(
+            repository_list_query, cursor_class=DictCursor
+        )
         results = result_set.fetchall()
+
         if len(results) == 0:
             raise ClickException(
-                f"Specified repository name {repo_name} not found in database {database} and schema {schema}"
+                f"Image repository '{repo_name}' does not exist in database '{self.get_database()}' and schema '{self.get_schema()}' or not authorized."
             )
-        else:
-            if len(results) > 1:
-                raise Exception(
-                    f"Found more than one repositories with name {repo_name}. This is unexpected."
-                )
+        elif len(results) > 1:
+            raise ClickException(
+                f"Found more than one image repository with name matching '{repo_name}'. This is unexpected."
+            )
+        return results[0]
 
-        return f"https://{results[0][4]}"
+    def get_repository_url(self, repo_name: str, with_scheme: bool = True):
+        if not is_valid_unquoted_identifier(repo_name):
+            raise ValueError(
+                f"repo_name '{repo_name}' is not a valid unquoted Snowflake identifier"
+            )
+        repo_row = self.get_repository_row(repo_name)
+        if with_scheme:
+            return f"https://{repo_row['repository_url']}"
+        else:
+            return repo_row["repository_url"]
 
     def get_repository_api_url(self, repo_url):
         """
@@ -60,3 +74,9 @@ class ImageRepositoryManager(SqlExecutionMixin):
         path = parsed_url.path
 
         return f"{scheme}://{host}/v2{path}"
+
+    def create(self, name: str):
+        try:
+            return self._execute_schema_query(f"create image repository {name}")
+        except ProgrammingError as e:
+            handle_object_already_exists(e, ObjectType.IMAGE_REPOSITORY, name)