+++ src/snowflake/cli/plugins/spcs/services/manager.py
@@ -1,10 +1,15 @@
 from pathlib import Path
 from typing import List, Optional
 
+from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
 from snowflake.cli.plugins.object.common import Tag
-from snowflake.cli.plugins.spcs.common import strip_empty_lines
+from snowflake.cli.plugins.spcs.common import (
+    handle_object_already_exists,
+    strip_empty_lines,
+)
 from snowflake.connector.cursor import SnowflakeCursor
+from snowflake.connector.errors import ProgrammingError
 
 
 class ServiceManager(SqlExecutionMixin):
@@ -24,7 +29,7 @@ class ServiceManager(SqlExecutionMixin):
         spec = self._read_yaml(spec_path)
 
         query = f"""\
-            CREATE SERVICE IF NOT EXISTS {service_name}
+            CREATE SERVICE {service_name}
             IN COMPUTE POOL {compute_pool}
             FROM SPECIFICATION $$
             {spec}
@@ -53,7 +58,10 @@ class ServiceManager(SqlExecutionMixin):
         if comment:
             query.append(f"COMMENT = {comment}")
 
-        return self._execute_schema_query(strip_empty_lines(query))
+        try:
+            return self._execute_schema_query(strip_empty_lines(query))
+        except ProgrammingError as e:
+            handle_object_already_exists(e, ObjectType.SERVICE, service_name)
 
     def _read_yaml(self, path: Path) -> str:
         # TODO(aivanou): Add validation towards schema