+++ test_external_plugins/failing_plugin/src/snowflakecli/test_plugins/failing_plugin/commands.py
@@ -1,0 +1,5 @@
+import typer
+
+app = typer.Typer()
+
+raise RuntimeError("Some error in plugin")