+++ tests_integration/testing_utils/assertions/test_result_assertions.py
@@ -40,3 +40,24 @@ def assert_that_result_is_successful_and_done_is_on_output(
     assert_that_result_is_successful(result)
     assert result.output is not None
     assert json.loads(result.output) == {"message": "Done"}
+
+
+def assert_that_result_is_successful_and_executed_successfully(
+    result: CommandResult, is_json: bool = False
+) -> None:
+    """
+    Checks that the command result is {"status": "Statement executed successfully"} as either json or text output.
+    """
+    assert_that_result_is_successful(result)
+    if is_json:
+        success_message = {"status": "Statement executed successfully."}
+        assert result.json is not None
+        if isinstance(result.json, dict):
+            assert result.json == success_message
+        else:
+            assert len(result.json) == 1
+            assert result.json[0] == success_message
+    else:
+        assert result.output is not None
+        assert "status" in result.output
+        assert "Statement executed successfully" in result.output